lambda_backend/requirements.txt
autopep8==1.5.6
certifi==2020.12.5
chardet==4.0.0
idna==2.10
numpy==1.20.2
opencv-python==4.5.1.48
Pillow==8.2.0
pycodestyle==2.7.0
requests==2.25.1
requests-toolbelt==0.9.1
toml==0.10.2
urllib3==1.26.4
